+++ src/tf/parse_expanddims.cpp
@@ -17,9 +17,9 @@ struct parse_expanddims : op_parser<parse_expanddims>
                           const tf_parser::node_info& info,
                           std::vector<instruction_ref> args) const
     {
-        std::vector<size_t> input_dims = args[0]->get_shape().lens();
+        std::vector<int> input_dims = args[0]->get_shape().lens();
         std::vector<int64_t> new_dims(input_dims.begin(), input_dims.end());
-        size_t num_dims = input_dims.size();
+        int num_dims = input_dims.size();
         int32_t dim     = args[1]->eval().at<int32_t>();
 
         if(dim < 0)
